app/src/test/java/org/technoserve/farmcollector/FarmCollectorAppTest.kt
package org.technoserve.farmcollector

/*
import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import androidx.work.Constraints
import androidx.work.ExistingPeriodicWorkPolicy
import androidx.work.NetworkType
import androidx.work.PeriodicWorkRequest
import androidx.work.WorkManager
import org.junit.Assert.*
import org.junit.Before
import org.junit.Rule
import org.junit.Test
import org.mockito.ArgumentMatchers.any
import org.mockito.ArgumentMatchers.anyString
import org.mockito.Mockito
import org.mockito.Mockito.mockStatic
import java.util.concurrent.TimeUnit

class FarmCollectorAppTest {

    @get:Rule
    val instantTaskExecutorRule = InstantTaskExecutorRule()

    private lateinit var mockWorkManager: WorkManager
    private lateinit var app: FarmCollectorApp

    @Before
    fun setup() {
        mockWorkManager = Mockito.mock(WorkManager::class.java)
        mockStatic(WorkManager::class.java).use { mockedStatic ->
            mockedStatic.`when`<Any> { WorkManager.getInstance(any()) }.thenReturn(mockWorkManager)
        }
        app = FarmCollectorApp()
    }

    @Test
    fun `test onCreate initializes WorkManager`() {
        // Arrange
        val expectedTag = "sync_work_tag"
        val expectedPolicy = ExistingPeriodicWorkPolicy.UPDATE

        // Act
        app.onCreate()

        // Assert
        Mockito.verify(mockWorkManager).enqueueUniquePeriodicWork(
            Mockito.eq(expectedTag),
            Mockito.eq(expectedPolicy),
            Mockito.any()
        )
    }

    @Test
    fun `test initializeWorkWorker creates PeriodicWorkRequest`() {
        // Arrange
        val expectedInterval = 2L
        val expectedTimeUnit = TimeUnit.HOURS

        // Act
        val workRequest = app.initializeWorkManager()

        // Assert
        Mockito.verify(mockWorkManager).enqueueUniquePeriodicWork(
            anyString(),
            Mockito.any(),
            Mockito.argThat { request: PeriodicWorkRequest ->
                request.workSpec.intervalDuration == TimeUnit.HOURS.toMillis(expectedInterval) &&
                        request.workSpec.constraints.requiredNetworkType == NetworkType.CONNECTED
            }
        )
    }

    @Test
    fun `test initializeWorkWorker throws exception when network not connected`() {
        // Arrange
        val mockConstraints = Constraints.Builder()
            .setRequiredNetworkType(NetworkType.NOT_REQUIRED)
            .build()
        Mockito.`when`(app.initializeWorkManager()).thenThrow(Exception("Required network type is not connected"))

        // Act & Assert
        try {
            app.initializeWorkManager()
            fail("Expected an exception to be thrown")
        } catch (exception: Exception) {
            assertTrue(exception.message?.contains("Required network type") == true)
        }
    }

}*/